requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiofiles==23.2.1 